refactor(structure): reorganize transforms into SurVAE-inspired hierarchy
Reshape `_src/` to mirror the SurVAE taxonomy the package is already built
around, with a secondary technique-based split inside each category. The
previous layout mixed axes: some files grouped by technique
(`conv.py`, `coupling.py`, `rotation.py`), some by SurVAE category
(`stochastic.py`, `surjections.py`), and one by domain (`periodic.py`).

New layout:

  _src/
    transforms/
      base.py, bijections/, surjections/, stochastic/
    bijections/
      coupling/, elementwise/, linear/, normalization/,
      reshape/, periodic/, classic/
    surjections/
      dimension/, abs.py, sort.py, pool.py
    stochastic/
      vae.py, permutation.py
    flows/
      survae.py, gaussianization.py, rbig.py
    inference/
      numpyro_compat.py, numpyro_guide.py, train.py
    nn/                                  # reserved for future local NN blocks

Public API at `gauss_flows.*` is unchanged: the 48 `__all__` entries (VAE,
SurVAEFlow, gaussianization_flow, entropy, FlowDist, FlowGuide, etc.) are
re-exported identically from the nested subpackages. Test imports continue
to work without modification.

Private imports inside `_src/` that referenced the old flat modules (e.g.
`gauss_flows._src.flows`, `gauss_flows._src.transforms.coupling`) now
point at the deeper leaves. `ConditionalDistribution` stays at
`_src/_protocols.py` and `AbstractSurjection` / `AbstractStochastic` /
`_IdentitySurjection` stay at `_src/transforms/base.py` so the
`tests/test_survae_base.py` private-path imports still resolve.

Verified with: full fast test suite (183 passed), `ruff check`,
`ruff format --check`, `ty check src/gauss_flows`.
