+++ src/gauss_flows/_src/flows/__init__.py
@@ -1,0 +1,21 @@
+"""Flow constructors and the :class:`SurVAEFlow` container.
+
+- :mod:`.survae` holds the bijection/surjection chain container.
+- :mod:`.gaussianization` holds the RBIG-style flow factories.
+- :mod:`.rbig` holds the classic iterative-RBIG constructor.
+"""
+
+from gauss_flows._src.flows.gaussianization import (
+    coupling_gaussianization_flow,
+    gaussianization_flow,
+)
+from gauss_flows._src.flows.rbig import iterative_rbig
+from gauss_flows._src.flows.survae import SurVAEFlow
+
+
+__all__ = [
+    "SurVAEFlow",
+    "coupling_gaussianization_flow",
+    "gaussianization_flow",
+    "iterative_rbig",
+]
