+++ src/gauss_flows/_src/inference/numpyro_compat.py
