+++ src/gauss_flows/_src/nn/__init__.py
@@ -1,0 +1,9 @@
+"""Reserved namespace for future local neural-network building blocks.
+
+All conditioner MLPs and learned embeddings currently come from
+``equinox`` and ``flowjax``. This subpackage is an empty placeholder — add
+project-local NN components here when a future transform needs one that
+isn't provided by those libraries.
+"""
+
+__all__: list[str] = []
