+++ src/gauss_flows/_src/transforms/bijections/coupling/spline.py
@@ -1,0 +1,61 @@
+"""Rational quadratic spline coupling layer."""
+
+from __future__ import annotations
+
+from typing import ClassVar
+
+import jax.numpy as jnp
+from flowjax.bijections import AbstractBijection, Coupling, RationalQuadraticSpline
+from jaxtyping import ArrayLike, PRNGKeyArray
+
+
+class RQSplineCoupling(AbstractBijection):
+    """Rational quadratic spline coupling layer.
+
+    Like AffineCoupling, but uses a rational quadratic spline as the transformer.
+
+    Args:
+        key: JAX random key.
+        shape: Shape of the input (n_dims,).
+        n_bins: Number of spline bins. Defaults to 8.
+        interval: Spline interval. Defaults to 5.0.
+        nn_width: Hidden layer width of the conditioner MLP. Defaults to 64.
+        nn_depth: Depth of the conditioner MLP. Defaults to 2.
+    """
+
+    shape: tuple[int, ...]
+    cond_shape: ClassVar[None] = None
+    _coupling: AbstractBijection
+
+    def __init__(
+        self,
+        key: PRNGKeyArray,
+        shape: tuple[int, ...],
+        n_bins: int = 8,
+        interval: float = 5.0,
+        nn_width: int = 64,
+        nn_depth: int = 2,
+    ):
+        if len(shape) != 1:
+            raise ValueError("RQSplineCoupling only supports 1D inputs.")
+        n_dims = shape[0]
+        self.shape = shape
+
+        spline = RationalQuadraticSpline(knots=n_bins, interval=interval)
+        self._coupling = Coupling(
+            key=key,
+            transformer=spline,
+            untransformed_dim=n_dims // 2,
+            dim=n_dims,
+            nn_width=nn_width,
+            nn_depth=nn_depth,
+        )
+
+    def transform_and_log_det(self, x: ArrayLike, condition=None):
+        return self._coupling.transform_and_log_det(jnp.asarray(x), condition)
+
+    def inverse_and_log_det(self, y: ArrayLike, condition=None):
+        return self._coupling.inverse_and_log_det(jnp.asarray(y), condition)
+
+
+__all__ = ["RQSplineCoupling"]
