+++ src/gauss_flows/_src/transforms/bijections/reshape/__init__.py
@@ -1,0 +1,6 @@
+"""Reshape / shape-rewriting bijections (volume-preserving)."""
+
+from gauss_flows._src.transforms.bijections.reshape.squeeze import Squeeze
+
+
+__all__ = ["Squeeze"]
