+++ src/gauss_flows/_src/transforms/bijections/reshape/squeeze.py
@@ -1,0 +1,50 @@
+"""Volume-preserving squeeze / reshape bijection."""
+
+from __future__ import annotations
+
+from typing import ClassVar
+
+import jax.numpy as jnp
+from flowjax.bijections import AbstractBijection
+from jaxtyping import ArrayLike
+
+
+class Squeeze(AbstractBijection):
+    """Squeeze operation for image-like inputs.
+
+    Reshapes spatial dimensions into channels, halving the spatial resolution
+    and quadrupling the channels. Commonly used in multi-scale flow architectures.
+
+    Args:
+        shape: Shape of the input (H, W, C) or (N, C).
+    """
+
+    shape: tuple[int, ...]
+    cond_shape: ClassVar[None] = None
+    out_shape: tuple[int, ...]
+
+    def __init__(self, shape: tuple[int, ...]):
+        self.shape = shape
+        if len(shape) == 1:
+            # 1D case: split into two halves → (n//2, 2)
+            n = shape[0]
+            if n % 2 != 0:
+                raise ValueError("1D squeeze requires even dimension size.")
+            self.out_shape = (n // 2, 2)
+        elif len(shape) == 2:
+            h, c = shape
+            if h % 2 != 0:
+                raise ValueError("Height must be divisible by 2 for Squeeze.")
+            self.out_shape = (h // 2, c * 2)
+        else:
+            self.out_shape = shape
+
+    def transform_and_log_det(self, x: ArrayLike, condition=None):
+        # Volume-preserving: log det = 0
+        return jnp.asarray(x).reshape(self.out_shape), jnp.zeros(())
+
+    def inverse_and_log_det(self, y: ArrayLike, condition=None):
+        return jnp.asarray(y).reshape(self.shape), jnp.zeros(())
+
+
+__all__ = ["Squeeze"]
