+++ src/gauss_flows/_src/transforms/surjections/dimension/__init__.py
@@ -1,0 +1,7 @@
+"""Dimensionality-changing surjections: Augment (grow) and Slice (shrink)."""
+
+from gauss_flows._src.transforms.surjections.dimension.augment import Augment
+from gauss_flows._src.transforms.surjections.dimension.slice import Slice
+
+
+__all__ = ["Augment", "Slice"]
